restructuring into cmd and pkg main folders
