+++ cmd/main.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"flag"
 	"github.com/op/go-logging"
-	"github.com/vimukthi-git/beanstalkg/architecture"
-	"github.com/vimukthi-git/beanstalkg/operation"
+	"github.com/vimukthi-git/beanstalkg/pkg/architecture"
+	"github.com/vimukthi-git/beanstalkg/pkg/operation"
 	"net"
 	"net/http"
 	_ "net/http/pprof"
