+++ pkg/architecture/mock_priority_queue.go
