+++ pkg/architecture/protocol.go
