+++ pkg/architecture/protocol_config.go
