+++ pkg/architecture/protocol_test.go
